requirements.txt
amqp==2.6.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==3.6.4.0
celery==4.4.7
Django==5.0.2
djangorestframework==3.14.0
flower==0.9.7
humanize==4.9.0
kombu==4.6.11
pillow==10.2.0
prometheus-client==0.8.0
pytz==2024.1
redis==5.0.2
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.10.0
vine==1.3.0
